updated typos
