+++ lib/nouns.js
@@ -1,0 +1,8 @@
+function Nouns(){
+  this.mirror = true;
+  this.random = true;
+  this.risk = true;
+  this.traffic = true;
+}
+
+module.export = Nouns;
