+++ lib/verbs.js
@@ -1,0 +1,8 @@
+function Verbs(){
+  this.doubt = true;
+  this.deserve = true;
+  this.recognise = true;
+  this.prevent = true;
+}
+
+module.exports = Verbs;
